+++ cantera18/branches/scons/Cantera/src/thermo/VPSSMgr_IdealGas.cpp
@@ -101,7 +101,7 @@ namespace Cantera {
 			   "ideal_gas: " + speciesNode.name());
       }
     }
-    if ((int) m_Vss.size() < k+1) {
+    if (m_Vss.size() < k+1) {
       m_Vss.resize(k+1, 0.0);
     }
